Log DB exceptions on SiteDBEditPage
SiteDBEditPage is smart, and rolls back failed DB calls, showing the user a "something went wrong" message. However the exception is just thrown away by default. This means we never get reports of failed queries, and makes development tricky. So instead always process the exception and continue so it gets outputted in development, and logged in production.