Remove setting teh site speed sample rate from the default commands.
Google's Tag Assistant Chrome Extension complains about the quoted int value. I'm fairly certain it doesn't break anything, but remove it to be on the safe side, as it doesn't lose a huge amount of information in our GA reports. We're going to replace all this code with Universal Analytics soon, and we can re-add the site speed sample rate then.