Add module for reporting data to WhiteChipmunk.com
WhiteChipmunk is a custom tracker being used by a subcontractor for one
of our clients.

This code is not perfect, but a very quick and dirty implementation for
a short run experiement that can be cleaned up if it needs to live for
very long.

Note that it does it's own cookie writing and reading - this is because
SiteCookieModule prefixes cookie names with the site's shortname, and we
don't want this behaviour for this tracker, so that multiple sites that
live in the same domain can read and write the same cookie. If this
lives for long enough, the prefixing of the cookie should be added to
SiteCookieModule as an option that can be turned off instead of
duplicating code here.