Prevent all dimensions from being re-processed.
Calling processManual() will in turn call prepareForProcessing(). That
method resets the image dimensions bindings to an empty collection.
Since that collection is noe empty but we are still in the loop all
further calls to hasDimensionBinding will be false which will cause all
dimensions to be processed instead of just the missing ones.