Remove all references to nate go search
Nate go search is no longer being used, and we don't want to bother
upgrading it to PHP8. So this removes all references to nate go search
from site. We'll be doing this for other packages too. We'll need a
majopr version upgrade since we're deprecating a number of classes as
well as number of addToSearchQueue methods.